requirements.txt
# Core setup
python-dotenv==1.1.0
setuptools==75.8.0
wheel==0.45.1

# OpenAI and embeddings
openai>=1.3.0
sentence-transformers==2.3.1
tqdm==4.66.2
tiktoken

# API & server
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
gunicorn==21.2.0

# Vector store & document parsing
faiss-cpu==1.7.4
pymupdf4llm~=0.0.21

# LangChain components
langchain-community>=0.0.20
langchain-core>=0.1.21

# Google Cloud integration
google-cloud-storage>=2.13.0

# Data science stack
pandas>=2.0.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
numpy>=1.24.0
chardet>=5.0.0

# Utility libs
requests>=2.31.0
typing-extensions>=4.7.0
xmltodict>=0.13.0
beautifulsoup4>=4.12.0
tenacity>=8.2.0
backoff

# Optional: other LLM providers
anthropic>=0.5.0

# Pydantic model parsing
pydantic==2.6.1

# Additional dependencies for enhanced functionality
aiofiles>=23.0.0
asyncio-pool>=0.6.0
regex>=2023.0.0
dataclasses-json>=0.6.0